+++ preferences.py
@@ -1,6 +1,9 @@
 import bpy
 import sys
+import os
+import subprocess
 import shutil
+import tempfile
 import zipfile
 from pathlib import Path
 import urllib.request
@@ -17,7 +20,6 @@ def get_addon_prefs():
     function to read current addon preferences properties
     access with : get_addon_prefs().super_special_option
     '''
-    import os 
     addon_name = os.path.splitext(__name__)[0]
     preferences = bpy.context.preferences
     addon_prefs = preferences.addons[addon_name].preferences
@@ -45,40 +47,143 @@ class SWD_OT_open_addon_prefs(bpy.types.Operator):
         open_addon_prefs()
         return {'FINISHED'}
 
+def get_ffmpeg_version(ffpath='ffmpeg'):
+
+    try:
+        ret = subprocess.run(args=[ffpath, '-version'], text=True, capture_output=True)
+    except Exception:
+        print(f'impossible to check ffmpeg version at : {ffpath}')
+        return 'Error checking version'
+
+    if ret.stderr:
+        ret = ret.stderr
+    else:
+        ret = ret.stdout
+
+    if get_addon_prefs().debug:
+        print()
+        print(f'Path: {ffpath}')
+        print('Version:')
+        print(ret)
+
+    line = ret.split('\n')[0]
+    if not 'version' in line.lower():
+        return 'Could not get version'
+    return line.split(' Copyright')[0]
+
 class SWD_OT_check_ffmpeg(bpy.types.Operator):
     """check if ffmpeg is in path"""
     bl_idname = "swd.check_ffmpeg"
-    bl_label = "Check ffmpeg in system path"
+    bl_label = "Check FFmpeg"
     bl_options = {'REGISTER', 'INTERNAL'}
 
     def invoke(self, context, event):
-        import  shutil
+        prefs = get_addon_prefs()
+        import shutil
         self.sys_path_ok = shutil.which('ffmpeg')
+
+        self.sys_path_version = None
+        self.ffbin = None
+        self.local_version = None
+
+        self.ffmpeg_at_path = None
+        self.ffmpeg_at_path_version = None
+
+        if prefs.path_to_ffmpeg and Path(prefs.path_to_ffmpeg).exists():
+            self.ffmpeg_at_path = prefs.path_to_ffmpeg
+            self.ffmpeg_at_path_version = get_ffmpeg_version(prefs.path_to_ffmpeg)
 
         # check for executable
         self.local_ffmpeg = False
-
-        is_windows_os = sys.platform.startswith('win')
-        if is_windows_os:
-            ffbin = Path(__file__).parent / 'ffmpeg.exe'
-            self.local_ffmpeg = ffbin.exists()
+        if sys.platform.startswith('win'):
+            self.ffbin = Path(__file__).parent / 'ffmpeg.exe'
+            self.local_ffmpeg = self.ffbin.exists()
         else:
             # Mac / linux
-            ffbin = Path(__file__).parent / 'ffmpeg'
-            self.local_ffmpeg = ffbin.exists()
+            self.ffbin = Path(__file__).parent / 'ffmpeg'
+            self.local_ffmpeg = self.ffbin.exists()
+        
+        if self.local_ffmpeg:
+            self.ffbin = str(self.ffbin)
+            self.local_version = get_ffmpeg_version(self.ffbin)
 
-        return context.window_manager.invoke_props_dialog(self, width=400)
+        if self.sys_path_ok:
+            self.sys_path_version = get_ffmpeg_version()
+
+        return context.window_manager.invoke_props_dialog(self, width=500)
 
     def draw(self, context):
         layout = self.layout
 
-        if self.local_ffmpeg:
-            layout.label(text='ffmpeg found in addon folder (this binary will be used)', icon='CHECKMARK')
+        if self.ffmpeg_at_path:
+            box = layout.box()
+            col = box.column()
+            col.label(text='This executable will be used', icon='INFO')
+            col.label(text='ffmpeg at specified path', icon='CHECKMARK')
+            col.label(text=self.ffmpeg_at_path, icon='FILE')
+            col.label(text=self.ffmpeg_at_path_version)
+
+        elif self.local_ffmpeg:
+            box = layout.box()
+            col = box.column()
+            col.label(text='This executable will be used', icon='INFO')
+            col.label(text='ffmpeg in addon folder', icon='CHECKMARK')
+            col.label(text=self.ffbin, icon='PLUGIN')
+            col.label(text=self.local_version)
 
         if self.sys_path_ok:
-            layout.label(text='ffmpeg is in system PATH', icon='CHECKMARK')
+            box = layout.box()
+            col = box.column()
+            if not self.ffmpeg_at_path and not self.local_ffmpeg:
+                col.label(text='This executable will be used', icon='INFO')
+            col.label(text='ffmpeg in system PATH', icon='CHECKMARK')
+            col.label(text=self.sys_path_ok, icon='SYSTEM')
+            col.label(text=self.sys_path_version)
+
         else:
             layout.label(text='ffmpeg is not in system PATH', icon='X') # CANCEL
+
+    def execute(self, context):
+        return {'FINISHED'}
+
+class SWD_OT_remove_ffmpeg(bpy.types.Operator):
+    """remove local ffmpeg in folder"""
+    bl_idname = "swd.remove_ffmpeg"
+    bl_label = "Remove ffmpeg"
+    bl_options = {'REGISTER', 'INTERNAL'}
+
+    def invoke(self, context, event):
+        # check for executable
+        self.local_ffmpeg = False
+        if sys.platform.startswith('win'):
+            self.ffbin = Path(__file__).parent / 'ffmpeg.exe'
+            self.local_ffmpeg = self.ffbin.exists()
+        else:
+            # Mac / linux
+            self.ffbin = Path(__file__).parent / 'ffmpeg'
+            self.local_ffmpeg = self.ffbin.exists()
+        
+        if self.local_ffmpeg:
+            try:
+                os.remove(self.ffbin)
+                self.mess = f'Local ffmpeg removed from addon folder'
+            except Exception as e:
+                print(e)
+                self.mess = f'Error: Could not remove local ffmpeg file'
+        else:
+            self.mess = 'No ffmpeg file found in addon folder'
+
+        return context.window_manager.invoke_props_dialog(self, width=700)
+
+    def draw(self, context):
+        layout = self.layout
+        layout.label(text='Executable path:')
+        layout.label(text=str(self.ffbin))
+        layout.label(text=self.mess)
+
+        if self.mess.startswith('Error'):
+            layout.label(text='Try removing ffmpeg executable manually')
+            layout.operator('wm.path_open', text='Open addon folder', icon='FILE_FOLDER').filepath = str(Path(__file__).parent)
 
     def execute(self, context):
         return {'FINISHED'}
@@ -190,13 +295,14 @@ class SWD_sound_waveform_display_addonpref(bpy.types.AddonPreferences):
         name="Waveform Height", description="How waveform size adapt to editor height", 
         default='RELATIVE', options={'HIDDEN', 'SKIP_SAVE'},
         items=(
-            ('RELATIVE', 'Relative', 'Waveform size is relative and adapt to editor hieght', 0),
+            ('RELATIVE', 'Relative', 'Waveform size is relative and adapt to editor height', 0),
             ('ABSOLUTE', 'Absolute', 'Waveform size is absolute and do not change according to editor size', 1),
             ))
 
     debug : BoolProperty(
-        name="Verbose Mode",
-        description="Verbose/Debug mode. Enable prints in console to follow script behavior",
+        name="Debug Mode",
+        description="Enable prints in console to follow script behavior\
+            \nkeep mixdown audio in temp folder for inspection",
         default=False, options={'HIDDEN'})
 
     show_help : BoolProperty(
@@ -217,16 +323,37 @@ class SWD_sound_waveform_display_addonpref(bpy.types.AddonPreferences):
         box.prop(self, "height_mode")
         box.prop(self, "debug")
 
+        if self.debug:
+            dbox = box.box()
+            # dbox.label(text='Debug mode add prints in console', icon='INFO')
+            col = dbox.column()
+            col.label(text='Debug tools:')
+            col.label(text='Audio mix and Waveform image are generated in operating system "temp" folder')
+            col.label(text='- Waveform image file is named ".tmp_scene_waveform.png"')
+            col.label(text='- Sound Mixdown file is named "tmp_scene_mixdown.wav"')
+            # col.label(text='While debug is enabled, audio file will not be removed from temp folder')
+            col.operator('wm.path_open', text='Open Temp Folder', icon='FILE_FOLDER').filepath = str(Path(tempfile.gettempdir()))
+            
+            col.separator()
+            col.label(text='When using ffmpeg auto-install below, ffmpeg executable is placed in addon folder')
+            col.label(text='If no waveform are generated you can try removing ffmpeg executable')
+            col.label(text='or point to another using below path')
+            col.operator('wm.path_open', text='Open addon folder', icon='FILE_FOLDER').filepath = str(Path(__file__).parent)
+            
+
         box = layout.box()
         box.label(text='FFmpeg check and auto-installation:')
+        
         col = box.column()
         # col.label(text="This addon use ffmpeg to generate the waveform (need a recent version)")
 
         col.label(text="This addon need an ffmpeg binary (Need to be installed if not in PATH)")
         row = col.row()
-        row.operator('swd.check_ffmpeg', text='Check If FFmpeg In PATH', icon='PLUGIN')
+        row.operator('swd.check_ffmpeg', text='Check if FFmpeg is ready', icon='PLUGIN')
         # row = col.row()
+        row = col.row(align=True)
         row.operator('swd.download_ffmpeg', text='Auto-install FFmpeg', icon='IMPORT')
+        row.operator('swd.remove_ffmpeg', text='Remove auto-installed', icon='X')
 
         # row.operator("wm.call_menu", text="Ffmpeg Manual Install Help", icon='HELP').name = "SWD_MT_help_ffmpeg_install"
         # row.operator('wm.url_open', text='FFmpeg Download Page', icon='URL').url = 'https://www.ffmpeg.org/download.html'
@@ -302,6 +429,7 @@ def help_infos(layout):
 classes=(
 SWD_OT_open_addon_prefs,
 SWD_OT_check_ffmpeg,
+SWD_OT_remove_ffmpeg,
 SWD_OT_download_ffmpeg,
 # SWD_MT_help_ffmpeg_install,
 SWD_sound_waveform_display_addonpref,