+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [1.0.2] - 2024-05-05
+
+### Added
+
+- More debug capabilities:
+    - new informations in preferences UI to resolve issues
+    - Operator to remove auto-installed ffmpeg
+
 ## [1.0.1] - 2024-03-01
 
 ### Fixed